Value return (#52)
* Fix handleGet

The lab manual states that the server sends a message in the format "VALUE <name> <value> (<upper>)\n" upon receiving a  valid GET request. 
Currently, the server only returns "VALUE <value> (<upper>)\n" -> the message is missing the name of the setting. This commit fixes that issue

* Added missing space

there was a space missing from the server message.