unreliable-chat-client: setting BURST-LEN through client now works
when trying to chance BURST-LEN within a client in the old code would not do anything, because the server does nothing in this case. I assume it was assumed that leaving case "BURST-LEN" empty would cause to do the same as DELAY-LEN but it does not. Now it does.

VU-net ID: lvo233